+++ laktory/models/datasinks/basedatasink.py
@@ -221,16 +221,6 @@ class BaseDataSink(BaseModel, PipelineChild):
     def serialize_path(self, value: Path) -> str:
         return value.as_posix()
 
-    @property
-    def upstream_node_names(self) -> list[str]:
-        """Pipeline node names required to write sink"""
-        return []
-
-    @property
-    def data_sources(self):
-        """Get all sources feeding the sink"""
-        return []
-
     # -------------------------------------------------------------------------------- #
     # CDC                                                                              #
     # -------------------------------------------------------------------------------- #
@@ -363,6 +353,9 @@ class BaseDataSink(BaseModel, PipelineChild):
         view_definition:
             View definition for table data sinks of `VIEW` type
         """
+
+        logger.info("Write initiated.")
+
         if getattr(self, "table_type", None) == "VIEW":
             if view_definition is None:
                 raise ValueError(f"`view_definition` for '{self._id}' is `None`")